gen-machineconf: Remove --bsp,--custom cmdline args
This patch will remove the --bsp and --custom commandline args instead
uses the --machine-name to specify the machine name and validate it with
machineconf.json file.

Rename gen_machine_file_name to machine_conf_file to be in sync with
plnx_conf_file variable.

Update the sysconfig(config) with machine-name specified with
--machine-name(commandline) to reflect in menuconfig/config.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>