gen_yocto_machine.py: Use machine json file
Remove the static dict for machine name instead use the json file.
If the given machine-name found in json treat it as BSP and generate the
machine conf file using the params defined(Ex: DT board file) in there
and include the generic machine based on the soc-family.
If it doesnot found in json treat it as custom and include generic
machine. Default machine names are
{microblaze,zynq,zynqmp,versal,versal-net}-generic and DT machine set to
'template'

Generate machine file as $PLATFORM + $device_id.conf if generic machine
used.

Set the DT board file based on the json file if machine name found.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>