gen_plnx_machine.py: Adjust IMAGE_FSTYPES override for initramdisk
The IMAGE_FSTYPES:pn-<ramdisk>:<soc> needs to use INITRAMFS_IMAGE as the PN
value.  (INITRAMFS_IMAGE is always set to the same value as the initramdisk.)

This avoids the warning:

  WARNING: Variable key IMAGE_FSTYPES:pn-${INITRAMFS_IMAGE} (${INITRAMFS_FSTYPES}) replaces original key IMAGE_FSTYPES:pn-petalinux-initramfs-image (${INITRAMFS_FSTYPES}).

The value is set at ${INITRAMFS_IMAGE} so that a user can override, in their
local.conf the INITRAMFS_IMAGE value with any ramdisk, not just the default
PetaLinux ramdisk.  PLNX configuration needs to match the same behavior to
avoid the warning.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>