processor.yaml: Moving configs from header to cfg
To align with upstream u-boot moving configs from
header file to config.
Removing CONFIG_CMD_IRQ config as the required support for
Microblaze got removed with the below
https://lore.kernel.org/all/5f2decc5a30a5678490ebde26d8c6f5a5f873cda.1654684731.git.michal.simek@amd.com/
Defining value to CONFIG_SYS_LONGHELP as if check for this got changed
without value it is treating as not defined.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-for-series: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>