Remove unused IMAGE_FSTYPES assignment
IMAGE_FSTYPES probably should not be set in the machine config as the
default values are sensible and they may need to be overwritten
elsewhere.

These values were being set in the Kria machine configs but have been
removed and dont seem to be present in any other machine configs so
should probably be removed from this script.

Signed-off-by: John Toomey <john.toomey@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>