hw-description.tcl: add proper check for axi_emc
For axi_emc controller we can connect flash or ddr
We use to check for bank type but that is not
giving proper result. Earlier we statically defined
MIG device so we have not seen this issue. Now removed
the static values seeing the issue.
Using the AXI_EMC_BOARD_INTERFACE config is giving proper
result.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>