xilinx_mirrors.py: file:// is appending extra
Adding extra file:// for sstate and downloads
when we are using XILINX_INT_SSTATES and XILINX_INT_DOWNLOADS
internal variables.
With extra file:// we can not use pointed sstate and downloads
properly so sstate rebuild will happen.

This extra appending only added when we install the tool
local as dep.conf file wont be there.

Note: This change is only for internal xilinx users

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>