Revert "gen_yocto_machine.py: Instead of ATF_CONSOLE_DEFAULT variable add ATF_CONSOLE"
This reverts commit 40291876fab0efc807059827ed7a3965ed9a08ea.

Unable to see ATF prints on uart0.
The rootcause of this issue is in 2023.1 we have changed all the ATF
name instances to TF-A and missed in one place where we will set
default serial console for TF-A to uart0 and default it was set to uart1.
But in 2023.1 and 2023.2 we are seeing ATF console prints on uart0
only as yocto hardcoded it to
uart0 using this variable "ATF_CONSOLE_DEFAULT".
Now we fixed the hardcoded issue and
then the original issue came out.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>