Kconfig.part: Removed TF-A CONFIG_SUBSYSTEM_PRELOADED_BL33_BASE entry
There are two menuconfigs for u-boot load address, one will be used to pass to TF-A and other is for u-boot.
As both are for the same purpose one config should be fine to avoid user errors
so, removing TF-A config and using the u-boot text base to pass to TF-A also.

Signed-off-by: John Vicky Vykuntapu <jvykunta@amd.com>
Acked-by: Varalaxmi Bingi <Varalaxmi.bingi@amd.com>