common_utils.py: Add new GetBitbakeVars function
Getting basic values from bitbake is a fairly common behavior.  Instead of
running bitbake -e ... in various places, centralize this with a new function.

Addtionally we can't run bitbake commands, unless we have bitbake in the path
so create a simple 'HaveBitbake' function to see if we have bitbake.

Replace GetBBPath() with a call to the new GetBitbakeVars function.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>