Make gen-machineconf capable of directly working with bitbake
Using the tinfoil interface into bitbake, we can do most actions directory.
This is much faster, as it avoids the fork/exec and bitbake setup overhead.

This does require that we shutdown the bitbake connection when we are done,
otherwise the system will hang waiting on a lock.  This is accomplished with
the new ShutdownBitbake function which is run prior to application exit OR
before we need to run another command that uses bitbake (such as
bitbake-layers.)

This series greatly reduced the initial setup time, and simplified the display
even when bitbake was required to build necessary tools.  If bitbake/tinfoil
are unavailable, we fall back to the previous behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>