Sync hardware info printing
Both SDT and XSCT flows will now print associated hardware information.
This has been moved to a debug print, as it isn't directly useful by the
end user.

Both SDT and XSCT now use a common function to print this data.

SDT move the setting of the default machine so it can be printed in the
debug.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>