update_buildconf.py: GenSdtConf use logger.plain for reference output
When outputting the local.conf fragment, check if conf_file was set by
the user.  (Don't assuming it's stdout any longer.)  If not set, then
we output using the 'logger.plain'.  Visually the user gets the same
results and we avoid writing directly to stdout.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series:Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>