update_buildconf.py: Always output local.conf info
Rename GenSdtConf to GenLocalConf.  This better reflects the new purpose of
this function.

Always output the MACHINE = value, the remaining values are dependent upon a
multiconfig target.  Note, it's possible to use multiconfigs with XSA, but
unlikely we will implement this.

The multiconfig is controlled by CONFIG_YOCTO_BBMC_...

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series:Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>