multiconfigs.py: Replace BASE_TMPDIR with MC_TMPDIR_PREFIX
In order to simplify the settings that the user needs, the BASE_TMPDIR
has been reworked.  The new MC_TMPDIR_PREFIX sets the prefix for the
generated multiconfig (and machine.conf) TMPDIR values.

The user can override this setting in their local.conf, or it will
automatically inherit from the system's TMPDIR.

At the same time the multiconfig log info as been cleaned up to present
a more unified format when getting multiconfig components.

The format is:
  Generating <processor> baremetal configuration ...

Finally the BBMULTICONFIG setting is now injected into the generated
machine .conf file.  This makes it easy to be regenerated and avoids
the user from having to make any changes to their local.conf file.
This enables the standard MACHINE=... bitbake ... behavior which will
activate the multiconfigs and build the proper configuration
automatically.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series:Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>