post_process_config.py: Use serial0 alias when only one serial is enabled
Currently, if only serial1 is enabled in the design, gen-machine-conf
sets the console to serial1. However, in the aliases, SDT maps serial0
to the serial1 node and assigns serial1 alias to Coresight.

This change fixes the inconsistency by checking the number of serial
IPs enabled in the design and setting the console as serial0 when only
one serial IP is present.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Acked-by: Varalaxmi Bingi <Varalaxmi.bingi@amd.com>