gen-machine-conf: Revise error message for missing components
The existing error messages have confused numerous users.

Make it clear that bitbake being missing is a problem, and try to give
the user a clue as to what is wrong.  (Note, this doesn't change behavior
if the user has explicitly disabled bitbake in some way.)

When searching for tools, make it clear that we will generate the
necessary tools for the user if bitbake is configured.  Give some
suggestions on how the configuration may be broken.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Reviewed-by: Sandeep Gundlupet Raju <Sandeep.gundlupet-raju@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>