common_utils.py: Rework RunCmd to provide more diagnostic
When a command fails, the generated exception may not always be useful.
Often it will simply output "[ERROR]" which doesn't tell the user
anything about why we stopped processing.

Instead, expand failure cases to include what command was run, the
environment, and make it clear if stdout and stderr have contents
or were blank.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>